pyproject.toml
[project]
name = "ontology-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=6.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-neo4j>=0.7.0",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "mcp>=1.26.0",
    "neo4j>=6.1.0",
    "openai>=2.30.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]
